player2Dの円移動完了
