+++ OculusGame/Assets/MyFolders/DaikiUmemura/player2D.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: b369f420ae5b5f042b08203bb320e475
+folderAsset: yes
+timeCreated: 1455526144
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
