--- OculusGame/Assets/MyFolders/RyoheiSuzuki/Player2DMove.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 701c6c392749ab44f91a81678a7b8a0c
-timeCreated: 1455499902
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
