+++ OculusGame/Assets/MyFolders/RyoheiSuzuki/player2d.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 623fe25f9a94ad5479e63f6ce52d0644
+folderAsset: yes
+timeCreated: 1455584322
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
