+++ OculusGame/Assets/MyFolders/RyoheiSuzuki/player2d/Player2D.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a833aa907ba1d6145a42b1a144841677
+timeCreated: 1455583638
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
